+++ schema/750$alert.queueOut.push.sql
@@ -1,9 +1,10 @@
-ALTER PROCEDURE [alert].[queue.push]
+ALTER PROCEDURE [alert].[queueOut.push]
     @port varchar(255),
     @channel varchar(128),
     @recipient [core].[arrayList] READONLY,
     @content nvarchar(max),
     @priority int = 0,
+    @messageInId BIGINT = 0,
     @meta [core].[metaDataTT] READONLY
 AS
 BEGIN
@@ -13,14 +14,14 @@ BEGIN
         DECLARE @actorId bigint = (select [auth.actorId] from @meta)
 
 		IF @actorId IS NULL
-			RAISERROR(N'alert.queue.push.missingCreatorId', 16, 1);
+			RAISERROR(N'alert.queueOut.push.missingCreatorId', 16, 1);
 
         SELECT 'inserted' resultSetName;
 
-        INSERT INTO [alert].[messageQueue](port, channel, recipient, content, createdBy, createdOn, statusId, priority)
+        INSERT INTO [alert].[messageOut](port, channel, recipient, content, createdBy, createdOn, statusId, priority, messageInId)
         OUTPUT INSERTED.id, INSERTED.port, INSERTED.channel, INSERTED.recipient, INSERTED.content, INSERTED.createdBy, INSERTED.createdOn,
-                @statusName as status, INSERTED.priority
-        SELECT @port, @channel, LTRIM(RTRIM([value])), @content, @actorId, SYSDATETIMEOFFSET(), @statusId, @priority
+                @statusName as status, INSERTED.priority, INSERTED.messageInId
+        SELECT @port, @channel, LTRIM(RTRIM([value])), @content, @actorId, SYSDATETIMEOFFSET(), @statusId, @priority, @messageInId
         FROM @recipient
     END TRY
     BEGIN CATCH